Johanna - Gate

+++ app/Http/Controllers/RoleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 use App\Permission\Models\Role;
 use App\Permission\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 
 use Illuminate\Http\Request;
@@ -16,6 +17,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('haveaccess','role.index');
         //
         $roles =  Role::orderBy('id','Desc')->paginate(2);
 
@@ -29,6 +31,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('haveaccess','role.create');
         //
         $permissions = Permission::get();
 
@@ -43,6 +46,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('haveaccess','role.create'); //aplique para el store y create
         //
         $request->validate([
             'name'          => 'required|max:50|unique:roles,name',
@@ -53,10 +57,10 @@ class RoleController extends Controller
         $role = Role::create($request->all());
         
         //Validacion 
-        if ($request->get('permission')) {
+        //if ($request->get('permission')) {
             //return $request->all();
             $role->permissions()->sync($request->get('permission'));
-        }
+        //}
         return redirect()->route('role.index')
             ->with('status_success','Role saved successfully'); 
     }
@@ -69,6 +73,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        //Gate::authorize('haveaccess','role.index');
+        $this->authorize('haveaccess','role.show'); //otra manera de acceder como gate 
         //
         $permission_role=[];
 
@@ -90,6 +96,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('haveaccess','role.edit');
         //
         $permission_role=[];
 
@@ -112,6 +119,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        $this->authorize('haveaccess','role.edit');
         //
         $request->validate([
             'name'          => 'required|max:50|unique:roles,name,'.$role->id,
@@ -121,10 +129,10 @@ class RoleController extends Controller
 
         $role->update($request->all());
         
-        if ($request->get('permission')) {
+        //if ($request->get('permission')) {
             //return $request->all();
             $role->permissions()->sync($request->get('permission'));
-        }
+        //}
         return redirect()->route('role.index')
             ->with('status_success','Role updated successfully'); 
     }
@@ -137,6 +145,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('haveaccess','role.destroy');
         //
         $role->delete();
